chore(Storybook): inline Babel config declaration in Storybook 'main' file
Prior to this commit, we tried to only consume Babel when using Storybook. That was done by moving
the 'babel.config.js' file to the '.storybook' directory. Though, when later updating to the
latest Storybook 7 version, it does no longer work. Hence we strive for using the SWC for Next, we
move the babel config declaration to Storybook's 'main' file.